Split page attributes mutations into multiple files
