+++ saleor/graphql/page/mutations/__init__.py
@@ -1,7 +1,11 @@
+from .page_attribute_assign import PageAttributeAssign
+from .page_attribute_unassign import PageAttributeUnassign
 from .page_create import PageCreate
 from .page_delete import PageDelete
+from .page_reorder_attribute_values import PageReorderAttributeValues
 from .page_type_create import PageTypeCreate
 from .page_type_delete import PageTypeDelete
+from .page_type_reorder_attributes import PageTypeReorderAttributes
 from .page_type_update import PageTypeUpdate
 from .page_update import PageUpdate
 
@@ -12,4 +16,8 @@ __all__ = [
     "PageTypeCreate",
     "PageTypeUpdate",
     "PageTypeDelete",
+    "PageAttributeAssign",
+    "PageAttributeUnassign",
+    "PageReorderAttributeValues",
+    "PageTypeReorderAttributes",
 ]
